Update to PyMC5
- removes dependency on Python 3.11
- adds CI tests for Python 3.12
- removes Theano dependency
